server/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
dnspython==2.7.0
idna==3.10
MarkupSafe==3.0.2
PyJWT==1.7.1
pymongo==4.11.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
